Add verify checksum job using coreutils hashsum
To make up for a number of limitations of using hashdeep utilities we
use this commit to implement the verify checksum jobs with the
hashsum utilities found in coreutils. Additional testing has been
added to the microservice as well.

Log files of the checksum output no longer exists, but additional
granular logging is now done in the microservice output and available
in the Archivematica logs themselves.

Checksums that are now supported include:

   * MD5, SHA1, SHA256, SHA512, Blake2

Functionality should be nearly equivalent to that of the previous
job but with the additional rigor needed to make sure checks are done
against all digital files.